+++ src/platforms/arduino_platform/arduino_systemtime.cpp
@@ -3,7 +3,7 @@
 
 int64_t NOW() {
 
-    static volatile uint32_t g_lastMicroseconds = 0;
+    static volatile int64_t g_lastMicroseconds = 0;
     static volatile int64_t g_currentTime = 0;
 
     uint32_t time = micros();
